Added support for /agent, /catalog and /health endpoints.
Stubs for tests for these endpoints added.
Not all methods are implemented and/or covered by tests. I've tried to update the readme to reflect what remains to be done.

Readme extended to cover the opts parameter as well as instructions on running tests and configuration files for the Consul agent used by the tests have been added.